Volume subscription new plan tests added and implemented
